main/requirements.txt
audioread==2.1.9
fastapi==0.68.0
google-cloud-firestore==2.5.2
httpx==0.23.0
python-multipart==0.0.5
uvicorn==0.15.0
